feat(md-annotations): improved support
Besides the #:: syntax, now it also supports #::[0-9] syntax to define a annotation number.
This annotation number can then be used to write a multi-line text for it a the end of the code block.

Example:
```javascript
let some = code #::1

```

This will render the annotation text at the end of the line with the content: `This is a annotation for the code above`.